+++ packages/ui/certd-server/src/plugins/plugin-lib/tencent/access.ts
@@ -87,7 +87,7 @@ export class TencentAccess extends BaseAccess {
 
   buildEndpoint(endpoint: string) {
     // 改成  xxx.intl.tencentcloudapi.com
-    return `${endpoint.replace(".tencentcloudapi.com", "")}${this.intlDomain()}tencentcloudapi.com`;
+    return `${endpoint.replace("tencentcloudapi.com", "")}${this.intlDomain()}tencentcloudapi.com`;
   }
 
   async getCallerIdentity() {